+++ backend/EasyMeets.Core/EasyMeets.Core.WebAPI/Controllers/AvailabilityController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace EasyMeets.Core.WebAPI.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("[controller]")]
     public class AvailabilityController : ControllerBase
@@ -17,7 +18,6 @@ namespace EasyMeets.Core.WebAPI.Controllers
             _availabilityService = availabilityService;
         }
         
-        [Authorize]
         [HttpGet("slot/{id}")]
         public async Task<ActionResult<AvailabilitySlotDto>> GetAvailabilitySlotById(long id)
         {
@@ -25,7 +25,6 @@ namespace EasyMeets.Core.WebAPI.Controllers
             return Ok(result);
         }
 
-        [Authorize]
         [HttpPost]
         public async Task<IActionResult> CreateAvailabilitySlot([FromBody] SaveAvailabilitySlotDto slotDto)
         {
@@ -33,7 +32,6 @@ namespace EasyMeets.Core.WebAPI.Controllers
             return Ok();
         } 
 
-        [Authorize]
         [HttpGet("{id}/{teamId?}")]
         public async Task<ActionResult<UserPersonalAndTeamSlotsDto>> GetUserPersonalAndTeamSlotsAsync(long id, long? teamId)
         {
@@ -41,21 +39,18 @@ namespace EasyMeets.Core.WebAPI.Controllers
             return Ok(availabilitySlots);
         }
 
-        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult<AvailabilitySlotDto>> UpdateAvailabilitySlot(long id, [FromBody] SaveAvailabilitySlotDto updateAvailabilityDto)
         {
             return Ok(await _availabilityService.UpdateAvailabilitySlot(id, updateAvailabilityDto));
         }
 
-        [Authorize]
         [HttpPost("enabling/{id}")]
         public async Task<ActionResult<bool>> UpdateAvailabilitySlotEnabling(long id)
         {
             return Ok(await _availabilityService.UpdateAvailabilitySlotEnablingAsync(id));
         }
 
-        [Authorize]
         [HttpDelete("{slotId}")]
         public async Task<IActionResult> DeleteAvailabilitySlot(int slotId)
         {
@@ -64,19 +59,20 @@ namespace EasyMeets.Core.WebAPI.Controllers
         }
 
         [HttpGet("byLink/{link}")]
+        [AllowAnonymous]
         public async Task<ActionResult<AvailabilitySlotDto?>> GetByLink(string link)
         {
             return Ok(await _availabilityService.GetByLink(link));
         }
 
         [HttpPut("{link}/externalSchedule")]
+        [AllowAnonymous]
         public async Task<IActionResult> UpdateScheduleExternally(string link, [FromBody] ScheduleDto scheduleDto)
         {
             await _availabilityService.UpdateScheduleExternally(link, scheduleDto);
             return Ok();
         }
         
-        [Authorize]
         [HttpGet("validateLink")]
         public async Task<ActionResult<bool>> ValidatePageLinkAsync(long? id, string slotLink)
         {
