+++ backend/EasyMeets.Core/EasyMeets.Core.WebAPI/Controllers/MeetingController.cs
@@ -35,6 +35,13 @@ namespace EasyMeets.Core.WebAPI.Controllers
             var teamMembers = await _teamService.GetTeamMembersOfCurrentUserAsync(teamId);
             return Ok(teamMembers);
         }
+        
+        [HttpGet("ordered-times/{slotId}")]
+        [AllowAnonymous]
+        public async Task<ActionResult<List<OrderedMeetingTimesDto>>> GetOrderedMeetingTimes(long slotId)
+        {
+            return Ok(await _meetingService.GetOrderedMeetingTimesAsync(slotId));
+        }
 
         [HttpPost]
         public async Task<ActionResult<SaveMeetingDto>> SaveNewMeeting([FromBody] SaveMeetingDto newMeetingDto)
