+++ frontend/src/app/core/services/new-meeting.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { INewMeeting } from '@core/models/INewMeeting';
 import { INewMeetingMember } from '@core/models/INewMeetingTeamMember';
+import { IOrderedMeetingTimes } from '@core/models/IOrderedMeetingTimes';
 
 import { HttpInternalService } from './http-internal.service';
 
@@ -19,5 +20,9 @@ export class NewMeetingService {
 
     public saveNewMeeting(data: INewMeeting) {
         return this.httpService.postRequest<INewMeeting>(`${this.routePrefix}`, data);
+    }
+
+    public getOrderedMeetingTimes(slotId: bigint) {
+        return this.httpService.getRequest<IOrderedMeetingTimes[]>(`${this.routePrefix}/ordered-times/${slotId}`);
     }
 }
